LPD-45929 Prevent NPE in case no custom meta tags structure, is caught because of testMissingCtCollectionId test failure

+++ modules/apps/layout/layout-seo-service/src/main/java/com/liferay/layout/seo/internal/upgrade/v3_0_0/LayoutSEOEntryCustomMetaTagUpgradeProcess.java
@@ -98,6 +98,10 @@ public class LayoutSEOEntryCustomMetaTagUpgradeProcess extends UpgradeProcess {
 							LayoutSEOEntry.class.getName()),
 						"custom-meta-tags");
 
+				if (ddmStructure == null) {
+					return;
+				}
+
 				_ddmStorageLinkLocalService.deleteStructureStorageLinks(
 					ddmStructure.getStructureId());
